+++ delay-message-sqlite/src/main/java/com/yhh/dto/DelayDto.java
@@ -23,7 +23,7 @@ public class DelayDto {
     /**
      * 延时时间戳[精准到秒级别]
      */
-    private Integer delayTime;
+    private Long delayTime;
 
     @Override
     public String toString() {
@@ -68,11 +68,11 @@ public class DelayDto {
         this.message = message;
     }
 
-    public Integer getDelayTime() {
+    public Long getDelayTime() {
         return delayTime;
     }
 
-    public void setDelayTime(Integer delayTime) {
+    public void setDelayTime(Long delayTime) {
         this.delayTime = delayTime;
     }
 }
